+++ iosTraining-ReeenTests/WeatherViewControllerTests.swift
@@ -1,5 +1,5 @@
 //
-//  iosTraining_ReeenTests.swift
+//  WeatherViewControllerTests.swift
 //  iosTraining-ReeenTests
 //
 //  Created by 高橋 蓮 on 2023/06/06.
@@ -18,7 +18,7 @@ extension WeatherData {
     }
 }
 
-final class iosTraining_ReeenTests: XCTestCase {
+final class WeatherViewControllerTests: XCTestCase {
     var weatherViewController: WeatherViewController!
 
     func testWeatherConditionImageViewIsSunny() {
@@ -69,39 +69,5 @@ final class iosTraining_ReeenTests: XCTestCase {
         let mimTempLabel = weatherViewController.minTemperatureLabel.text
         XCTAssertEqual(maxTempLabel, "20")
         XCTAssertEqual(mimTempLabel, "10")
-    }
-
-    func testEncodeRequestParameters() {
-        let encoder = WeatherEncoder()
-        let date = DateComponents(calendar: Calendar(identifier: .gregorian), timeZone: TimeZone(secondsFromGMT: 0), year: 2023, month: 7, day: 7, hour: 12, minute: 1, second: 1).date!
-        let dummyRequestParameters = WeatherInformationRequest(area: "tokyo", date: date)
-        let expectData = #"{"area":"tokyo","date":"2023-07-07T12:01:01Z"}"#
-
-        do {
-            let encodedRequest = try encoder.encodeRequestParameters(dummyRequestParameters)
-            XCTAssertEqual(encodedRequest, expectData)
-        } catch {
-            XCTFail("Encoding failed with error: \(error)")
-        }
-    }
-
-
-    func testDecodeWeatherInfo() {
-        let decoder = WeatherDecoder()
-        let dummyWeatherInfo = """
-            {
-                "weather_condition": "sunny",
-                "max_temperature": 30,
-                "min_temperature": 20
-            }
-            """
-        do {
-            let weatherData = try decoder.decodeWeatherInfo(dummyWeatherInfo)
-            XCTAssertEqual(weatherData.weatherCondition, "sunny")
-            XCTAssertEqual(weatherData.maxTemperature, 30)
-            XCTAssertEqual(weatherData.minTemperature, 20)
-        } catch {
-            XCTFail("Decoding failed with error: \(error)")
-        }
     }
 }
